+++ src/commands/ban.ts
@@ -4,8 +4,8 @@ import { Message, Collection } from 'discord.js';
 export default class BanCommand implements Command {
     info = {
         names: ['ban'],
-        description: 'bannin crazy',
-        usage: 'ban'
+        description: 'bannin crazy / doesnt workk',
+        usage: '&ban'
     }
 
     async run(message: Message, args: string[], messages: any): Promise<void> {