+++ src/commands/eval.ts
@@ -6,7 +6,7 @@ export default class EvalCommand implements Command {
     info = {
         names: ['eval'],
         description: 'tylko dla rzadu',
-        usage: 'eval (code)'
+        usage: '&eval (code)'
     }
 
     set: any = {};