+++ src/commands/help.ts
@@ -6,7 +6,7 @@ export default class HelpCommand implements Command {
     info = {
         names: ['help'],
         description: 'Shows help',
-        usage: 'help'
+        usage: '&help'
     }
 
     run(message: Message, args: string[], messages: any): void {