+++ src/commands/tag.ts
@@ -6,7 +6,7 @@ export default class TagCommand implements Command {
     info = {
         names: ['tag'],
         description: 'taggo',
-        usage: 'tag ( title ) { description.. }'
+        usage: '&tag (title) {description..}'
     }
 
     async run(message: Message, args: string[], messages: any): Promise<any> {