+++ src/commands/weather.ts
@@ -6,7 +6,7 @@ export default class WeatherCommand implements Command {
     info = {
         names: ['weather'],
         description: 'chmurki or sloneczko',
-        usage: 'weather (city)'
+        usage: '&weather (city)'
     }
 
     run(message: Message, args: string[], messages: any): void {